backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.9.2
python-nmap==0.7.1
jinja2==3.1.4
paramiko==3.5.0
python-multipart==0.0.12